+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
-
-* TODO: Describe initial release.
+* Initial release.(ios still developing)
+## 0.0.2
+* Seperate TakePhotocallback. 
+## 0.0.3
+* Use different Image Plugin. 
+## 0.0.4
+* Asynchronous